user_service/requirements.txt
Package                Version
---------------------- -------
alembic                1.10.3
blinker                1.6
click                  8.1.3
Flask                  2.2.3
Flask-Login            0.6.2
Flask-Migrate          4.0.4
Flask-SQLAlchemy       3.0.3
itsdangerous           2.1.2
Jinja2                 3.1.2
Mako                   1.2.4
MarkupSafe             2.1.2
mysql-connector-python 8.0.32
passlib                1.7.4
pip                    23.0.1
protobuf               3.20.3
python-dotenv          1.0.0
setuptools             65.5.0
SQLAlchemy             2.0.9
typing_extensions      4.5.0
Werkzeug               2.2.3
wheel                  0.37.1
